fix: bad error message
ChildNotFound error had an invalid error message.
Don't output NotFound messages for ptpl.

+++ io-engine/src/bdev/nexus/nexus_share.rs
@@ -255,7 +255,9 @@ impl<'n> From<&Nexus<'n>> for NexusPtpl {
 impl PtplFileOps for NexusPtpl {
     fn destroy(&self) -> Result<(), std::io::Error> {
         if let Some(path) = self.path() {
-            std::fs::remove_file(path)?;
+            if path.exists() {
+                std::fs::remove_file(path)?;
+            }
         }
         Ok(())
     }
